don't export native audio package to web
